Issue 629 - Don't timeout connections while waiting for auth
Previously, during server auth, there are no packets sent back and forth between the client and server, which causes the client to timeout after 20 seconds (assuming this is in case the server has shutdown/crashed/etc). As reported in issue 629, this may not be long enough. I ended up tracing this issue back to the how the server sends CMD_ALIVE packets every 10 seconds through sendAll which only sends to authenticated clients. So I created a new keepAliveAll() which sends a CMD_ALIVE packet to all cients regardless of auth state. In addition, I added in logic to calculate the ClientInfo's ping. I didn't add any logic to do anything with the ping.