Fix size computation
UASTC have two notions of sizes - the 'native' one, which is used by the compressed format and is always at least a multiple of four, and the 'original' one, which can be any size.

Converting to an RGBA image requires the 'original' size, and the 'native' one is used when using compressed formats.