[libs] Update dtoa_milo, reverting #48
The bounds check added in #48 caused #54 in Clang, and now appears
to also cause it in GCC. The result formats floating point numbers
with a fixed decimal in the wrong place, breaking game state
logging in EmptyEpsilon.

Update dtoa_milo to the current commit in dtoa-benchmarks, which
also reverts the bounds check.