Return on only first RenderLayer::onPointerDown
RenderLayer::onPointerDown checks every Renderable at a position
and passes the event to all of them, ignoring any return values.
This causes clicks to pass through one Renderable to others below
it.

For example, if the Lua console is open on a crew screen or GM
screen, clicking on the console's text field also passes the click
to elements beneath it, such as the radar. If you select an entity
on the GM screen radar, and then to click on the Lua console to
enter a script acting on getGMSelection(), the console text field
click unexpectedly passes through the console to the radar and
clears or changes the selection.

Instead, respect the Renderable's return value and return true if
the Renderable returns true.